Maruti_Divekar/Hackathon_Submission/requirements.txt
# Core UI Framework
gradio>=4.0.0

# HTTP and Web Scraping
requests>=2.28.0
beautifulsoup4>=4.12.0

# PDF Processing
pymupdf>=1.23.0

# Configuration Management
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Machine Learning and Vector Operations
scikit-learn>=1.3.0
numpy>=1.24.0
faiss-cpu>=1.7.4

# LLM Integration
openai>=1.0.0
tiktoken>=0.5.1

# Testing and Code Quality
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.5.0

# Optional Dependencies
langchain>=0.1.0  # Optional: Advanced LLM orchestration
langchain-community>=0.0.10  # Optional: Community integrations
